"Project submit once"
Answer:
For autocomplete("The", "movies.txt", 10), the function takes 0.26 seconds to match prefix query. In particular, the slowest part is queue$push(), it takes 0.20 seconds to run, which is 76.92% of overall time. The way to improve priority queue is to try other queue packages or implement priority queue during trie insertion instead of autocomplete() function.